split constant / adaptive brgihtness modules for Split classes
